Incorporate review comments except three items
