+++ Bus Scheduler/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import Home from './pages/home/index';
+import Home from './features/home/index';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
