+++ Bus Scheduler/src/services/enums.ts
@@ -1,0 +1,4 @@
+export enum ApiErrorMessages {
+    NotFound = 'The requested resource was not found.',
+    Unauthorized = 'You do not have the necessary permissions.'
+}
