+++ Bus Scheduler/src/shared/components/tripSearch/store/enums.ts
@@ -1,0 +1,3 @@
+export enum ActionKind {
+    'LIST_TRIP' = 'LIST_TRIP'
+}
