+++ Bus Scheduler/src/shared/components/tripSearch/store/types.ts
@@ -1,4 +1,4 @@
-import { BaseContextDataType } from "../../../store/types";
+import { BaseContextDataType } from "../../../../store/types";
 
 export type TripListType =
   {
@@ -19,6 +19,4 @@ export type TripSearchActionsType = {
   listTrips: (payload: BaseContextDataType<TripListType>) => void
 }
 
-export enum ActionKind {
-  'LIST_TRIP' = 'LIST_TRIP'
-}
+
