+++ Bus Scheduler/src/store/types.ts
@@ -1,6 +1,5 @@
 
 import { ReactNode, Context } from 'react'
-import { ActionKind } from '../components/tripSearch/store/types'
 
 
 export type BaseContextDataType<T> = {
@@ -8,7 +7,7 @@ export type BaseContextDataType<T> = {
 }
 
 export type ActionType<T> = {
-    type: ActionKind, payload: BaseContextDataType<T>
+    type: string, payload: BaseContextDataType<T>
 }
 
 export type BaseContextType<T> = BaseContextDataType<T> & {
