+++ Bus Scheduler/vite.config.ts
@@ -8,13 +8,14 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, './src'),
-      pages: `${path.resolve(__dirname, './src/containers/')}`,
-      components: `${path.resolve(__dirname, './src/components/')}`,
-      configs: `${path.resolve(__dirname, './src/configs/')}`,
-      assets: `${path.resolve(__dirname, './src/assets/')}`,
-      services: `${path.resolve(__dirname, './src/services/')}`,
-      store: `${path.resolve(__dirname, './src/store/')}`,
-      utils: `${path.resolve(__dirname, './src/utils/')}`,
+      '@/features': `${path.resolve(__dirname, './src/features/')}`,
+      '@/components': `${path.resolve(__dirname, './src/components/')}`,
+      '@/configs': `${path.resolve(__dirname, './src/configs/')}`,
+      '@/assets': `${path.resolve(__dirname, './src/assets/')}`,
+      '@/services': `${path.resolve(__dirname, './src/services/')}`,
+      '@/store': `${path.resolve(__dirname, './src/store/')}`,
+      '@/utils': `${path.resolve(__dirname, './src/utils/')}`,
+      '@/shared': `${path.resolve(__dirname, './src/shared/')}`,
     },
   },
 })
